Refactor color bar classes
