+++ bokeh/models/annotations/legends.py
@@ -27,6 +27,7 @@ from ...core.enums import (
     LegendLocation,
     Orientation,
 )
+from ...core.has_props import abstract
 from ...core.properties import (
     Auto,
     Bool,
@@ -74,12 +75,11 @@ __all__ = (
 # General API
 #-----------------------------------------------------------------------------
 
-class ColorBar(Annotation):
-    ''' Render a color bar based on a color mapper.
-
-    See :ref:`userguide_annotations_color_bars` for information on plotting color bars.
-
-    '''
+@abstract
+class BaseColorBar(Annotation):
+    # explicit __init__ to support Init signatures
+    def __init__(self, *args, **kwargs) -> None:
+        super().__init__(*args, **kwargs)
 
     location = Either(Enum(Anchor), Tuple(Float, Float), default="top_right", help="""
     The location where the color bar should draw itself. It's either one of
@@ -91,10 +91,6 @@ class ColorBar(Annotation):
         If the color bar is placed in a side panel, the location will likely
         have to be set to `(0,0)`.
     """)
-
-    # explicit __init__ to support Init signatures
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
 
     orientation = Either(Enum(Orientation), Auto, default="auto", help="""
     Whether the color bar should be oriented vertically or horizontally.
@@ -143,17 +139,6 @@ class ColorBar(Annotation):
 
     major_label_policy = Instance(LabelingPolicy, default=InstanceDefault(NoOverlap), help="""
     Allows to filter out labels, e.g. declutter labels to avoid overlap.
-    """)
-
-    color_mapper = Instance(ColorMapper, help="""
-    A color mapper containing a color palette to render.
-
-    .. warning::
-        If the `low` and `high` attributes of the ``ColorMapper`` aren't set, ticks
-        and tick labels won't be rendered. Additionally, if a ``LogTicker`` is
-        passed to the `ticker` argument and either or both of the logarithms
-        of `low` and `high` values of the color_mapper are non-numeric
-        (i.e. `low=0`), the tick and tick labels won't be rendered.
     """)
 
     margin = Int(30, help="""
@@ -226,14 +211,37 @@ class ColorBar(Annotation):
 
     background_fill_alpha = Override(default=0.95)
 
-class ContourColorBar(ColorBar):
+
+class ColorBar(BaseColorBar):
+    ''' Render a color bar based on a color mapper.
+
+    See :ref:`userguide_annotations_color_bars` for information on plotting color bars.
+
+    '''
     # explicit __init__ to support Init signatures
     def __init__(self, *args, **kwargs) -> None:
         super().__init__(*args, **kwargs)
 
-    #line_props = Include(LineProps)
-    line_renderer = Instance(GlyphRenderer)  # Maybe just want the line visuals?
+    color_mapper = Instance(ColorMapper, help="""
+    A color mapper containing a color palette to render.
+
+    .. warning::
+        If the `low` and `high` attributes of the ``ColorMapper`` aren't set, ticks
+        and tick labels won't be rendered. Additionally, if a ``LogTicker`` is
+        passed to the `ticker` argument and either or both of the logarithms
+        of `low` and `high` values of the color_mapper are non-numeric
+        (i.e. `low=0`), the tick and tick labels won't be rendered.
+    """)
+
+
+class ContourColorBar(BaseColorBar):
+    # explicit __init__ to support Init signatures
+    def __init__(self, *args, **kwargs) -> None:
+        super().__init__(*args, **kwargs)
+
     fill_renderer = Instance(GlyphRenderer)
+
+    line_renderer = Instance(GlyphRenderer)
 
 
 class LegendItem(Model):
