+++ bokeh/models/contour_renderer.py
@@ -77,18 +77,11 @@ class ContourRenderer(DataRenderer):
     fill_renderer = Instance(GlyphRenderer, default=_DEFAULT_CONTOUR_FILL_RENDERER, help="""
     """)
 
-    data = Dict(String, ColumnData(keys_type=String, values_type=Any))
-
-    # Do not need these if subclass ContourColorBar correctly?  Do need ticker?
-    color_mapper = Instance(ColorMapper)
     ticker = Instance(Ticker)
 
     def __setattr__(self, name: str, value: Unknown) -> None:
-        super().__setattr__(name, value)
-
         if name == "data":
-            # Should check these are set first.
-            # Copy across line/fill/hatch properties?
+            # Copy across line/fill/hatch properties.
             if "fill_data" in value:
                 new_fill_data = value["fill_data"]
                 old_fill_data = self.fill_renderer.data_source.data
@@ -104,14 +97,15 @@ class ContourRenderer(DataRenderer):
                     if name not in ("xs", "ys"):
                         new_line_data[name] = old_line_data[name]
                 self.line_renderer.data_source.data = new_line_data
+        else:
+            super().__setattr__(name, value)
 
     def color_bar(self) -> ContourColorBar:
-        #### Should accept other kwargs and pass them through?????
+        # Should accept other kwargs and pass them through?
         return ContourColorBar(
-            color_mapper=self.color_mapper,
             ticker=self.ticker,
             fill_renderer=self.fill_renderer,
-            line_renderer=self.line_renderer,   # Just send the glyph????  Or the line props?
+            line_renderer=self.line_renderer,
         )
 
 #-----------------------------------------------------------------------------
