+++ bokehjs/src/lib/models/annotations/contour_color_bar.ts
@@ -1,16 +1,14 @@
+import {BaseColorBar, BaseColorBarView} from "./base_color_bar"
+import {MultiLineView} from "../glyphs/multi_line"
+import {MultiPolygonsView} from "../glyphs/multi_polygons"
+import {GlyphRenderer, GlyphRendererView} from "../renderers/glyph_renderer"
 import {build_view} from "core/build_views"
 import * as p from "core/properties"
 import {BBox} from "core/util/bbox"
 import {Context2d} from "core/util/canvas"
-import {ColorBar, ColorBarView} from "./color_bar"
 
-import {GlyphRenderer, GlyphRendererView} from "../renderers/glyph_renderer"
-import {MultiLineView} from "../glyphs/multi_line"
-import {MultiPolygonsView} from "../glyphs/multi_polygons"
-
-export class ContourColorBarView extends ColorBarView {
+export class ContourColorBarView extends BaseColorBarView {
   override model: ContourColorBar
-  //override visuals: ContourColorBar.Visuals
 
   protected _fill_view: GlyphRendererView
   protected _line_view: GlyphRendererView
@@ -29,16 +27,8 @@ export class ContourColorBarView extends ColorBarView {
     super.remove()
   }
 
-  //override connect_signals(): void {
-  //  super.connect_signals()
-  //}
-
-  protected _paint_impl(ctx: Context2d, bbox: BBox): void {
-    console.log("_paint_impl", ctx, bbox)
-    //const {x, y, width, height} = bbox
-    //console.log("ticker", this._ticker)
-
-    // Need to check this first
+  protected _paint_colors(ctx: Context2d, bbox: BBox): void {
+    // Need to check this might be null
     const multi_polygons = this._fill_view.glyph as MultiPolygonsView
     ctx.save()
     const nfill = multi_polygons.data_size
@@ -51,7 +41,7 @@ export class ContourColorBarView extends ColorBarView {
     }
     ctx.restore()
 
-    // Need to check this first
+    // Need to check this might be null
     const multi_line = this._line_view.glyph as MultiLineView
     ctx.save()
     const nlines = multi_line.data_size
@@ -65,33 +55,12 @@ export class ContourColorBarView extends ColorBarView {
     }
     ctx.restore()
   }
-
-  protected override _render(): void {
-    // From parent...
-    const {ctx} = this.layer
-    ctx.save()
-    this._paint_bbox(ctx, this._inner_layout.bbox)
-
-    //this._paint_image(ctx, this._inner_layout.center_panel.bbox)
-    this._paint_impl(ctx, this._inner_layout.center_panel.bbox)
-
-    this._title_view.render()
-    this._axis_view.render()
-    ctx.restore()
-  }
-
-  /*override serializable_state(): SerializableState {
-    const {children = [], ...state} = super.serializable_state()
-    children.push(this._title_view.serializable_state())
-    children.push(this._axis_view.serializable_state())
-    return {...state, children}
-  }*/
 }
 
 export namespace ContourColorBar {
   export type Attrs = p.AttrsOf<Props>
 
-  export type Props = ColorBar.Props & {
+  export type Props = BaseColorBar.Props & {
     fill_renderer: p.Property<GlyphRenderer>
     line_renderer: p.Property<GlyphRenderer>
   }
@@ -99,7 +68,7 @@ export namespace ContourColorBar {
 
 export interface ContourColorBar extends ContourColorBar.Attrs {}
 
-export class ContourColorBar extends ColorBar {
+export class ContourColorBar extends BaseColorBar {
   override properties: ContourColorBar.Props
   override __view_type__: ContourColorBarView
