+++ bokehjs/src/lib/models/annotations/index.ts
@@ -5,6 +5,7 @@ export {OpenHead}          from "./arrow_head"
 export {NormalHead}        from "./arrow_head"
 export {TeeHead}           from "./arrow_head"
 export {VeeHead}           from "./arrow_head"
+export {BaseColorBar}      from "./base_color_bar"
 export {Band}              from "./band"
 export {BoxAnnotation}     from "./box_annotation"
 export {ColorBar}          from "./color_bar"
